workshops/Protein_Language_Model_Training/scripts/training/peft/requirements.txt

transformers==4.31.0
datasets==2.14.2
evaluate==0.4.0
torchinfo==1.8.0
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.41.0
scikit-learn==1.3.0
